refactor: Vertex 구조체를 Gfx로 이동
